requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
defusedxml==0.7.1
Django==2.2.14
django-allauth==0.42.0
django-anymail==7.2.1
django-bootstrap4==2.2.0
django-thumbs-v2==0.4.1
fonttools==4.28.5
fpdf==1.7.2
idna==3.3
importlib-metadata==1.7.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
Pillow==7.2.0
placeholder-pics==1.0.3
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
webcolors==1.11.1
zipp==3.6.0
